refactor(daemon): extract session end recovery
